finish authrization task

+++ api/urls/urls.controllers.js
@@ -1,8 +1,8 @@
 const Url = require("../../models/Url");
 const shortid = require("shortid");
 const User = require("../../models/User");
-
 const baseUrl = "http:localhost:8000/urls";
+
 
 exports.shorten = async (req, res) => {
   // create url code
@@ -10,7 +10,7 @@ exports.shorten = async (req, res) => {
   try {
     req.body.shortUrl = `${baseUrl}/${urlCode}`;
     req.body.urlCode = urlCode;
-    req.body.userId = req.params.userId;
+    req.body.userId = req.User._id;
     const newUrl = await Url.create(req.body);
     await User.findByIdAndUpdate(req.params.userId, {
       $push: { urls: newUrl._id },
@@ -36,13 +36,22 @@ exports.redirect = async (req, res) => {
 
 exports.deleteUrl = async (req, res) => {
   try {
+    const user = await User.findById(req.User._id);
     const url = await Url.findOne({ urlCode: req.params.code });
-    if (url) {
-      await Url.findByIdAndDelete(url._id);
-      return res.status(201).json("Deleted");
-    } else {
-      return res.status(404).json("No URL Found");
+    if(url){
+      if(user._id.equals(url.user)){
+        await Url.findByIdAndDelete(url._id);
+        return res.status(201).json("Deleted");
+      } else {
+        return res.status(404).json("No URL Found");
+      }
+      }
+
     }
+   
+   
+    if (url) {
+    
   } catch (err) {
     next(err);
   }
